requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt
gunicorn==20.0.4
idna==2.10
Markdown==3.3.4
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
urllib3==1.26.3
wget==3.2
whitenoise==5.2.0
